fix: 게시글 수정 오류 수정

+++ src/main/java/com/example/capstone/community/service/PostService.java
@@ -110,19 +110,24 @@ public class PostService {
                 .orElseThrow(() -> new EntityNotFoundException("Post Not Found"));
 
         // 기존 이미지 S3 삭제
-        List<String> imageUriList = objectMapper.readValue(post.getImageUris(), new TypeReference<>() {});
-        for (String imageUri : imageUriList) {
-            imageService.deleteImage(imageUri);
+        if (post.getImageUris() != null && !post.getImageUris().isEmpty()) {
+            List<String> imageUriList = objectMapper.readValue(post.getImageUris(), new TypeReference<>() {});
+            for (String imageUri : imageUriList) {
+                imageService.deleteImage(imageUri);
+            }
         }
 
-        List<String> newImageUriList = new ArrayList<>();
-        for (MultipartFile postImage : postImages) {
-            String imageUri = imageService.imageUpload(postImage, POST_IMAGE_DIR);
-            newImageUriList.add(imageUri);
-        }
+        String newImageUris = null;
+        if (postImages != null && !postImages.isEmpty()) {
+            List<String> newImageUriList = new ArrayList<>();
+            for (MultipartFile postImage : postImages) {
+                String imageUri = imageService.imageUpload(postImage, POST_IMAGE_DIR);
+                newImageUriList.add(imageUri);
+            }
 
-        // Json 배열 직렬화
-        String newImageUris = objectMapper.writeValueAsString(newImageUriList);
+            // Json 배열 직렬화
+            newImageUris = objectMapper.writeValueAsString(newImageUriList);
+        }
 
         post.updatePost(createPostReqDto.getTitle(), createPostReqDto.getContent(), newImageUris, createPostReqDto.getCity(), createPostReqDto.getProvince());
     }
